V3-1839 web satellite auth cookie subdomain differentiation
